+++ CHANGELOG.md
@@ -3,4 +3,7 @@
 ## 0.2.0
 - Added Docs
 ## 0.3.0
-- Fixed bugs occured during khalti payment success
+- Fixed bugs occurred during khalti payment success
+## 0.4.0
+- Temporarily removed feature to pass custom data while payment cause of issues
+- Added Option to choose Khalti payments Types (EBanking, SCT, Connect-IPS,Mobile Banking)